moved waitroom for game after tour
