+++ levels/stopgo-interactive/game/phantom.js
@@ -1,0 +1,19 @@
+/**
+ * # Phantom type implementation of the game stages
+ * Copyright(c) 2016 brenste <myemail>
+ * MIT Licensed
+ *
+ * http://www.nodegame.org
+ * ---
+ */
+var ngc = require('nodegame-client');
+var stepRules = ngc.stepRules;
+var constants = ngc.constants;
+
+module.exports = function(game, settings, treatmentName, gameRoom) {
+
+    // Must implement the stages here.
+
+    // ...
+
+};
